[core, lua, yaml] Port Job to YAML based enum

+++ src/map/latent_effect_container.cpp
@@ -820,7 +820,7 @@ auto CLatentEffectContainer::ProcessLatentEffect(CLatentEffect& latentEffect, bo
             expression = m_POwner->health.tp > latentEffect.GetConditionsValue();
             break;
         case xi::Latent::Subjob:
-            expression = m_POwner->GetSJob() == latentEffect.GetConditionsValue();
+            expression = static_cast<uint8>(m_POwner->GetSJob()) == latentEffect.GetConditionsValue();
             break;
         case xi::Latent::PetId:
             expression =
@@ -955,7 +955,7 @@ auto CLatentEffectContainer::ProcessLatentEffect(CLatentEffect& latentEffect, bo
                 {
                     if (member->id != m_POwner->id)
                     {
-                        if (member->GetMJob() == latentEffect.GetConditionsValue())
+                        if (static_cast<uint8>(member->GetMJob()) == latentEffect.GetConditionsValue())
                         {
                             expression = true;
                             break;
@@ -973,7 +973,7 @@ auto CLatentEffectContainer::ProcessLatentEffect(CLatentEffect& latentEffect, bo
 
                 for (auto* trust : leader->PTrusts)
                 {
-                    if (trust->GetMJob() == latentEffect.GetConditionsValue())
+                    if (static_cast<uint8>(trust->GetMJob()) == latentEffect.GetConditionsValue())
                     {
                         expression = true;
                         break;
@@ -1323,7 +1323,7 @@ auto CLatentEffectContainer::ProcessLatentEffect(CLatentEffect& latentEffect, bo
             }
             break;
         case xi::Latent::Mainjob:
-            expression = m_POwner->GetMJob() == latentEffect.GetConditionsValue();
+            expression = static_cast<uint8>(m_POwner->GetMJob()) == latentEffect.GetConditionsValue();
             break;
         case xi::Latent::EquippedInSlot:
             expression = latentEffect.GetSlot() == latentEffect.GetConditionsValue();
