+++ src/map/merit.cpp
@@ -422,7 +422,7 @@ int32 CMeritPoints::GetMeritValue(MERIT_TYPE merit, CCharEntity* PChar)
 
     if (PMerit)
     {
-        if (PMerit->catid < 5 || (PMerit->jobs & (1 << (PChar->GetMJob() - 1)) && PChar->GetMLevel() >= 75))
+        if (PMerit->catid < 5 || (PMerit->jobs & (1 << (static_cast<uint8>(PChar->GetMJob()) - 1)) && PChar->GetMLevel() >= 75))
         {
             meritValue = merit == MERIT_MAX_MERIT ? PMerit->count : std::min(PMerit->count, cap[PChar->GetMLevel()]);
         }
