+++ src/map/packets/char_sync.cpp
@@ -57,6 +57,6 @@ CCharSyncPacket::CCharSyncPacket(CCharEntity* PChar)
         ref<uint32>(0x1C)                           = CustomProperties[1]; // Noble Chocobo
     }
 
-    ref<uint8>(0x25) = PChar->jobs.job[PChar->GetMJob()];
+    ref<uint8>(0x25) = PChar->jobs.job[static_cast<uint8>(PChar->GetMJob())];
     ref<uint8>(0x27) = PChar->profile.mhflag & 0x20 ? 1 : 0; // MogExpansionFlag - Is 2nd floor unlocked.
 }
